+++ lib/std/optionals.nim
@@ -1,7 +1,7 @@
-#
+#Optional
 #
 #            Nim's Runtime Library
-#        (c) Copyright 2015 Nim Contributors
+#        (c) Copyright 2022 Nim Contributors
 #
 #    See the file "copying.txt", included in this
 #    distribution, for details about the copyright.
@@ -10,7 +10,7 @@
 ##[
 This module implements types which encapsulate an optional value.
 
-A value of type `Option[T]` either contains a value `x` (represented as
+A value of type `Optional[T]` either contains a value `x` (represented as
 `some(x)`) or is empty (`none(T)`).
 
 This can be useful when you have a value that can be present or not. The
@@ -26,7 +26,7 @@ in a string.
 ]##
 
 runnableExamples:
-  proc find(haystack: string, needle: char): Option[int] =
+  proc find(haystack: string, needle: char): Optional[int] =
     for i, c in haystack:
       if c == needle:
         return some(i)
@@ -41,7 +41,7 @@ The `get` operation demonstrated above returns the underlying value, or
 raises `UnpackDefect` if there is no value. Note that `UnpackDefect`
 inherits from `system.Defect` and should therefore never be caught.
 Instead, rely on checking if the option contains a value with the
-`isSome <#isSome,Option[T]>`_ and `isNone <#isNone,Option[T]>`_ procs.
+`isSome <#isSome,Optional[T]>`_ and `isNone <#isNone,Optional[T]>`_ procs.
 
 
 Pattern matching
@@ -50,7 +50,7 @@ Pattern matching
 .. note:: This requires the [fusion](https://github.com/nim-lang/fusion) package.
 
 [fusion/matching](https://nim-lang.github.io/fusion/src/fusion/matching.html)
-supports pattern matching on `Option`s, with the `Some(<pattern>)` and
+supports pattern matching on `Optional`s, with the `Some(<pattern>)` and
 `None()` patterns.
 
 .. code-block:: nim
@@ -75,28 +75,24 @@ when defined(nimPreviewSlimSystem):
   import std/assertions
 
 
-when (NimMajor, NimMinor) >= (1, 1):
-  type
-    SomePointer = ref | ptr | pointer | proc
-else:
-  type
-    SomePointer = ref | ptr | pointer
-
 type
-  Option*[T] = object
+  SomePointer = ref | ptr | pointer | proc # xxx what about cstring? proc vs closure?
+  Optional*[T] = object
     ## An optional type that may or may not contain a value of type `T`.
-    ## When `T` is a a pointer type (`ptr`, `pointer`, `ref` or `proc`),
-    ## `none(T)` is represented as `nil`.
-    when T is SomePointer:
-      val: T
-    else:
-      val: T
-      has: bool
+    val: T
+    has: bool
 
   UnpackDefect* = object of Defect
-  UnpackError* {.deprecated: "See corresponding Defect".} = UnpackDefect
 
-proc option*[T](val: sink T): Option[T] {.inline.} =
+template maybeOption*[T](_: typedesc[T]): untyped =
+  ## Return `T` if T is `ref | ptr | pointer | proc`, else `Optional[T]`
+  runnableExamples:
+    assert maybeOption(ref int) is ref int
+    assert maybeOption(int) is Optional[int]
+  when T is SomePointer: T
+  else: Optional[T]
+
+proc option*[T](val: sink T): Optional[T] {.inline.} =
   ## Can be used to convert a pointer type (`ptr`, `pointer`, `ref` or `proc`) to an option type.
   ## It converts `nil` to `none(T)`. When `T` is no pointer type, this is equivalent to `some(val)`.
   ##
@@ -106,89 +102,81 @@ proc option*[T](val: sink T): Option[T] {.inline.} =
   runnableExamples:
     type
       Foo = ref object
-        a: int
-        b: string
-
     assert option[Foo](nil).isNone
+    assert option(Foo(nil)).isNone
+    assert some(Foo(nil)).isSome
     assert option(42).isSome
-
-  result.val = val
-  when T isnot SomePointer:
+  when T is SomePointer:
+    result.val = val
+    result.has = val != nil
+  else:
+    result.val = val
     result.has = true
 
-proc some*[T](val: sink T): Option[T] {.inline.} =
-  ## Returns an `Option` that has the value `val`.
+proc some*[T](val: sink T): Optional[T] {.inline.} =
+  ## Returns an `Optional` that has the value `val`.
   ##
   ## **See also:**
   ## * `option proc <#option,T>`_
   ## * `none proc <#none,typedesc>`_
-  ## * `isSome proc <#isSome,Option[T]>`_
+  ## * `isSome proc <#isSome,Optional[T]>`_
   runnableExamples:
     let a = some("abc")
-
     assert a.isSome
     assert a.get == "abc"
 
-  when T is SomePointer:
-    assert not val.isNil
-    result.val = val
-  else:
-    result.has = true
-    result.val = val
+    let b: ref int = nil
+    assert some(b).isSome
+    assert option(b).isNone
 
-proc none*(T: typedesc): Option[T] {.inline.} =
-  ## Returns an `Option` for this type that has no value.
+  result.has = true
+  result.val = val
+
+proc none*(T: typedesc): Optional[T] {.inline.} =
+  ## Returns an `Optional` for this type that has no value.
   ##
   ## **See also:**
   ## * `option proc <#option,T>`_
   ## * `some proc <#some,T>`_
-  ## * `isNone proc <#isNone,Option[T]>`_
+  ## * `isNone proc <#isNone,Optional[T]>`_
   runnableExamples:
     assert none(int).isNone
 
   # the default is the none type
   discard
 
-proc none*[T]: Option[T] {.inline.} =
+proc none*[T]: Optional[T] {.inline.} =
   ## Alias for `none(T) <#none,typedesc>`_.
   none(T)
 
-proc isSome*[T](self: Option[T]): bool {.inline.} =
-  ## Checks if an `Option` contains a value.
+proc isSome*[T](self: Optional[T]): bool {.inline.} =
+  ## Checks if an `Optional` contains a value.
   ##
   ## **See also:**
-  ## * `isNone proc <#isNone,Option[T]>`_
+  ## * `isNone proc <#isNone,Optional[T]>`_
   ## * `some proc <#some,T>`_
   runnableExamples:
     assert some(42).isSome
     assert not none(string).isSome
+  self.has
 
-  when T is SomePointer:
-    not self.val.isNil
-  else:
-    self.has
-
-proc isNone*[T](self: Option[T]): bool {.inline.} =
-  ## Checks if an `Option` is empty.
+proc isNone*[T](self: Optional[T]): bool {.inline.} =
+  ## Checks if an `Optional` is empty.
   ##
   ## **See also:**
-  ## * `isSome proc <#isSome,Option[T]>`_
+  ## * `isSome proc <#isSome,Optional[T]>`_
   ## * `none proc <#none,typedesc>`_
   runnableExamples:
     assert not some(42).isNone
     assert none(string).isNone
+  not self.has
 
-  when T is SomePointer:
-    self.val.isNil
-  else:
-    not self.has
-
-proc get*[T](self: Option[T]): lent T {.inline.} =
-  ## Returns the content of an `Option`. If it has no value,
+proc get*[T](self: Optional[T]): lent T {.inline.} =
+  ## Returns the content of an `Optional`. If it has no value,
   ## an `UnpackDefect` exception is raised.
   ##
   ## **See also:**
-  ## * `get proc <#get,Option[T],T>`_ with a default return value
+  ## * `get proc <#get,Optional[T],T>`_ with a default return value
   runnableExamples:
     assert some(42).get == 42
     doAssertRaises(UnpackDefect):
@@ -198,9 +186,9 @@ proc get*[T](self: Option[T]): lent T {.inline.} =
     raise newException(UnpackDefect, "Can't obtain a value from a `none`")
   result = self.val
 
-proc get*[T](self: Option[T], otherwise: T): T {.inline.} =
-  ## Returns the content of the `Option` or `otherwise` if
-  ## the `Option` has no value.
+proc get*[T](self: Optional[T], otherwise: T): T {.inline.} =
+  ## Returns the content of the `Optional` or `otherwise` if
+  ## the `Optional` has no value.
   runnableExamples:
     assert some(42).get(9999) == 42
     assert none(int).get(9999) == 9999
@@ -210,8 +198,8 @@ proc get*[T](self: Option[T], otherwise: T): T {.inline.} =
   else:
     otherwise
 
-proc get*[T](self: var Option[T]): var T {.inline.} =
-  ## Returns the content of the `var Option` mutably. If it has no value,
+proc get*[T](self: var Optional[T]): var T {.inline.} =
+  ## Returns the content of the `var Optional` mutably. If it has no value,
   ## an `UnpackDefect` exception is raised.
   runnableExamples:
     var
@@ -226,11 +214,11 @@ proc get*[T](self: var Option[T]): var T {.inline.} =
     raise newException(UnpackDefect, "Can't obtain a value from a `none`")
   return self.val
 
-proc map*[T](self: Option[T], callback: proc (input: T)) {.inline.} =
-  ## Applies a `callback` function to the value of the `Option`, if it has one.
+proc map*[T](self: Optional[T], callback: proc (input: T)) {.inline.} =
+  ## Applies a `callback` function to the value of the `Optional`, if it has one.
   ##
   ## **See also:**
-  ## * `map proc <#map,Option[T],proc(T)_2>`_ for a version with a callback
+  ## * `map proc <#map,Optional[T],proc(T)_2>`_ for a version with a callback
   ##   which returns a value
   runnableExamples:
     var d = 0
@@ -245,16 +233,16 @@ proc map*[T](self: Option[T], callback: proc (input: T)) {.inline.} =
   if self.isSome:
     callback(self.val)
 
-proc map*[T, R](self: Option[T], callback: proc (input: T): R): Option[R] {.inline.} =
-  ## Applies a `callback` function to the value of the `Option` and returns an
-  ## `Option` containing the new value.
+proc map*[T, R](self: Optional[T], callback: proc (input: T): R): Optional[R] {.inline.} =
+  ## Applies a `callback` function to the value of the `Optional` and returns an
+  ## `Optional` containing the new value.
   ##
-  ## If the `Option` has no value, `none(R)` will be returned.
+  ## If the `Optional` has no value, `none(R)` will be returned.
   ##
   ## **See also:**
-  ## * `map proc <#map,Option[T],proc(T)>`_
-  ## * `flatMap proc <#flatMap,Option[T],proc(T)>`_ for a version with a
-  ##   callback that returns an `Option`
+  ## * `map proc <#map,Optional[T],proc(T)>`_
+  ## * `flatMap proc <#flatMap,Optional[T],proc(T)>`_ for a version with a
+  ##   callback that returns an `Optional`
   runnableExamples:
     proc isEven(x: int): bool =
       x mod 2 == 0
@@ -267,35 +255,35 @@ proc map*[T, R](self: Option[T], callback: proc (input: T): R): Option[R] {.inli
   else:
     none(R)
 
-proc flatten*[T](self: Option[Option[T]]): Option[T] {.inline.} =
-  ## Remove one level of structure in a nested `Option`.
+proc flatten*[T](self: Optional[Optional[T]]): Optional[T] {.inline.} =
+  ## Remove one level of structure in a nested `Optional`.
   ##
   ## **See also:**
-  ## * `flatMap proc <#flatMap,Option[T],proc(T)>`_
+  ## * `flatMap proc <#flatMap,Optional[T],proc(T)>`_
   runnableExamples:
     assert flatten(some(some(42))) == some(42)
-    assert flatten(none(Option[int])) == none(int)
+    assert flatten(none(Optional[int])) == none(int)
 
   if self.isSome:
     self.val
   else:
     none(T)
 
-proc flatMap*[T, R](self: Option[T],
-                    callback: proc (input: T): Option[R]): Option[R] {.inline.} =
-  ## Applies a `callback` function to the value of the `Option` and returns the new value.
+proc flatMap*[T, R](self: Optional[T],
+                    callback: proc (input: T): Optional[R]): Optional[R] {.inline.} =
+  ## Applies a `callback` function to the value of the `Optional` and returns the new value.
   ##
-  ## If the `Option` has no value, `none(R)` will be returned.
+  ## If the `Optional` has no value, `none(R)` will be returned.
   ##
   ## This is similar to `map`, with the difference that the `callback` returns an
-  ## `Option`, not a raw value. This allows multiple procs with a
-  ## signature of `A -> Option[B]` to be chained together.
+  ## `Optional`, not a raw value. This allows multiple procs with a
+  ## signature of `A -> Optional[B]` to be chained together.
   ##
   ## See also:
-  ## * `flatten proc <#flatten,Option[Option[A]]>`_
-  ## * `filter proc <#filter,Option[T],proc(T)>`_
+  ## * `flatten proc <#flatten,Optional[Optional[A]]>`_
+  ## * `filter proc <#filter,Optional[T],proc(T)>`_
   runnableExamples:
-    proc doublePositives(x: int): Option[int] =
+    proc doublePositives(x: int): Optional[int] =
       if x > 0:
         some(2 * x)
       else:
@@ -307,14 +295,14 @@ proc flatMap*[T, R](self: Option[T],
 
   map(self, callback).flatten()
 
-proc filter*[T](self: Option[T], callback: proc (input: T): bool): Option[T] {.inline.} =
-  ## Applies a `callback` to the value of the `Option`.
+proc filter*[T](self: Optional[T], callback: proc (input: T): bool): Optional[T] {.inline.} =
+  ## Applies a `callback` to the value of the `Optional`.
   ##
   ## If the `callback` returns `true`, the option is returned as `some`.
   ## If it returns `false`, it is returned as `none`.
   ##
   ## **See also:**
-  ## * `flatMap proc <#flatMap,Option[A],proc(A)>`_
+  ## * `flatMap proc <#flatMap,Optional[A],proc(A)>`_
   runnableExamples:
     proc isEven(x: int): bool =
       x mod 2 == 0
@@ -328,8 +316,8 @@ proc filter*[T](self: Option[T], callback: proc (input: T): bool): Option[T] {.i
   else:
     self
 
-proc `==`*[T](a, b: Option[T]): bool {.inline.} =
-  ## Returns `true` if both `Option`s are `none`,
+proc `==`*[T](a, b: Optional[T]): bool {.inline.} =
+  ## Returns `true` if both `Optional`s are `none`,
   ## or if they are both `some` and have equal values.
   runnableExamples:
     let
@@ -342,13 +330,10 @@ proc `==`*[T](a, b: Option[T]): bool {.inline.} =
     assert b == d
     assert not (a == b)
 
-  when T is SomePointer:
-    a.val == b.val
-  else:
-    (a.isSome and b.isSome and a.val == b.val) or (a.isNone and b.isNone)
+  (a.isSome and b.isSome and a.val == b.val) or (a.isNone and b.isNone)
 
-proc `$`*[T](self: Option[T]): string =
-  ## Get the string representation of the `Option`.
+proc `$`*[T](self: Optional[T]): string =
+  ## Get the string representation of the `Optional`.
   runnableExamples:
     assert $some(42) == "some(42)"
     assert $none(int) == "none(int)"
@@ -366,11 +351,11 @@ proc `$`*[T](self: Option[T]): string =
     else:
       result = "none(" & name(T) & ")"
 
-proc unsafeGet*[T](self: Option[T]): lent T {.inline.}=
+proc unsafeGet*[T](self: Optional[T]): lent T {.inline.}=
   ## Returns the value of a `some`. The behavior is undefined for `none`.
   ##
   ## **Note:** Use this only when you are **absolutely sure** the value is present
-  ## (e.g. after checking with `isSome <#isSome,Option[T]>`_).
-  ## Generally, using the `get proc <#get,Option[T]>`_ is preferred.
+  ## (e.g. after checking with `isSome <#isSome,Optional[T]>`_).
+  ## Generally, using the `get proc <#get,Optional[T]>`_ is preferred.
   assert self.isSome
   result = self.val